+++ ebong/src/app/component/LetterList/LetterListHooks.ts
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/react-query";
+import { letterClientApi } from "@/lib/client-api/letters";
+
+export const useLetterList = () => {
+  return useQuery({
+    queryKey: ["letters"],
+    queryFn: letterClientApi.getLetters,
+  });
+};
